Time: 6 ms (67.26%) | Memory: 25 MB (83.19%) - LeetSync

+++ 2713-find-the-divisibility-array-of-a-string/find-the-divisibility-array-of-a-string.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     vector<int> divisibilityArray(string word, int m) {
         vector<int> ans(word.size(), 0);
-        long long t = 0;
+        long t = 0;
         for(int i = 0; i < word.size(); ++i){
             t = (t * 10 + (word[i] - 48) ) % m;
             if(t == 0) ans[i] = 1;
